Define FAB delegate.
Created a delegate protocol that can be expanded upon to include a
bunch of additional events related to the behavior of the FAB. Right
now it only contains a function that notifies the delegate when a user
has tapped on a FAB that does not contain any defined sub items. Good
for FABs that represent a single parent action and don’t have nested
child actions.